package: remove non-conventional prefixes/suffixes from Github-fetched packages
On Github, a large number of projects name their tag
<some-prefix>-0.3-<some-suffix> (i.e release-3.0, poco-0.1-release,
etc.).

In most packages, we encode those prefix/suffix in the <pkg>_VERSION
variable.

The problem with this approach is that when used in conjunction with
release-monitoring.org, it doesn't work very well, because
release-monitoring.org has the concept of "version prefix/suffix" and
using that they drop the prefix/suffix to really get the version. For
example on https://release-monitoring.org/project/5418/ the latest
release of "poco" is "1.8.1", not "poco-1.8.1-release".

Therefore, a number of packages in Buildroot have a version that
doesn't match with release-monitoring.org.

Since really the version number of 1.8.1, is makes sense to update our
packages to drop these prefixes/suffixes.

This commit addreses the case of github-fetched packages with
non-conventional prefixes/suffixes.

Note that these changes modify the name of the files stored in DL_DIR,
which means that this will force a re-download of those package source
code for all users, and requires a change to their .hash file.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>