qt5: rework installation logic
Currently the installation of all Qt5 modules is a bit non-standard:

 - For the staging installation, we do just "make install" without any
   DESTDIR pointing to $(STAGING_DIR), because Qt has already
   hardcoded in the generated Makefiles the $(STAGING_DIR) location.

 - For the target installation, we manually copy the libraries, fonts,
   Qml files, plugins and other files from staging to target. Indeed,
   we can't use "make install" because "make install" will anyway
   install in $(STAGING_DIR) as it is hardcoded in the Makefiles.

This has a number of drawbacks:

 - We have to keep a lot of logic for the target installation on all
   qt5 .mk files, instead of just using "make install"

 - This target installation logic has lots of conditions: depending on
   whether QML is supported or not, whether examples are enabled or
   not, whether static linking is used or not, whether we're using Qt
   5.6 or the latest version, etc. This has already caused numerous
   build failures in the past, or files being missing on the target.

 - The current situation is inherently incompatible with per-package
   sysroot. Indeed, the Makefile generated by qmake hardcode the
   STAGING_DIR, and it's the STAGING_DIR that was used for the qt5base
   package. Therefore, when one builds for example qt5sensors, the
   "make install" installs in the per-package sysroot of qt5base
   instead of the one of qt5sensors..

For all those reasons, we want to fix up the installation logic so
that we can use the INSTALL_ROOT variable that is provided by
qmake-generated Makefiles, but is currently unusable due to
STAGING_DIR being hardcoded.

A first approach that we tried is to use the '-extprefix /usr'
argument to Qt's configure script. Using this, we got the following
results:

 - The installation of target components had a correct destination of
   $(INSTALL_ROOT)/usr/lib, $(INSTALL_ROOT)/usr/include, which allowed
   using INSTALL_ROOT as expected.

 - The installation of host components (qmake, moc, etc.) however was
   using $(INSTALL_ROOT)/$(HOST_DIR), which leads to host tools being
   installed without the STAGING_DIR/TARGET_DIR. We could fix this by
   moving them around back to the host directory.

 - Once that was fixed, unfortunately, for other packages than
   qt5base, the build failed because qmake was passing
   -I/usr/include/QtCore.

So we gave up on this approach, and took an approach similar to what
OpenEmbedded is doing: fixing up the Makefiles generated by qmake.

 - We replace all instances of $(INSTALL_ROOT)$(STAGING_DIR) by just
   $(INSTALL_ROOT), so that files aimed at going to the target are
   installed directly inside $(INSTALL_ROOT), and by using
   INSTALL_ROOT=$(TARGET_DIR) or INSTALL_ROOT=$(STAGING_DIR) we can
   install the target files properly.

 - We replace all instances of $(INSTALL_ROOT)$(HOST_DIR) by just
   $(HOST_DIR), so that files aimed at going on the host machine
   remained installed in $(HOST_DIR), and not in the concatenation of
   $(INSTALL_ROOT) and $(HOST_DIR).

Both of these actions are done in QT5_FIXUP_MAKEFILES that is called
at the end of the build step of all Qt5 packages.

Despite this, the qt5base package still causes an issue with files
installed under $(INSTALL_ROOT)/$(HOST_DIR), and so those files are
moved around back into $(HOST_DIR).

While these hacks are not that nice, they allow to drastically reduce
the amount of makefile logic needed to handle the installation of all
the Qt5 packages.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>