package/pkg-meson: ensure the global cross-compilation.conf file is correct
Currently, the cross-compilation.conf installed in
$(HOST_DIR)/etc/meson/cross-compilation.conf for use by the SDK is
generated in a post-install-staging hook of the toolchain package.

With per-package directory support enabled, this means that the
generated cross-compilation.conf contains references to the
per-package directory of the toolchain/ package, which is not want we
want. So instead, we generate this file in
TOOLCHAIN_TARGET_FINALIZE_HOOKS, so that the global paths are used.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>