package/genimage: add patch to support bootable flag in GPT tables
This commit adds a patch, submitted upstream, to add support for
setting the bootable flag in GPT partition tables. While we normally
don't accept feature patches in Buildroot, this one is really minimal
(two lines of code) and is in fact necessary for the STM32MP157
Discovery Kit defconfig.

Indeed, the STM32MP157 BootROM loads the first stage bootloader from
the GPT partitions named fsbl*, which requires a GPT partition
table. But then, the "U-Boot Generic Distro" mechanism identifies the
GPT partition that contains the extlinux.conf file to use thanks to
the "bootable" flag of GPT partitions.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>