package: drop use of GNU_HOST_NAME where possible
configure scripts are able to automatically detect the --build
argument, by running the config.guess script, but we override it with
--build=$(GNU_HOST_NAME).

It turns out that GNU_HOST_NAME is derived by Buildroot by running
config.guess, in the exact same version as the one that would be used
by the package itself, because we anyway overwrite any config.guess
provided by the package itself.

Using GNU_HOST_NAME is therefore useless in these situations, so we
simply remove such references. It will help in the effort to move the
gnuconfig handling from support/gnuconfig to a proper package.

After this commit, there are only two references to GNU_HOST_NAME left
in the tree:

 - In moarvm, which doesn't use autotools at all, and doesn't have any
   config.guess/config.sub logic.

 - In libnspr, which is autotools based, but needs GNU_HOST_NAME to be
   passed as the --host variable, due to libnspr build system being
   somewhat broken when it comes to cross-compilation.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>