package: change CONFIG_UPDATE_HOOK to use host-gnuconfig
Instead of using the config.guess/config.sub files from
support/gnuconfig, let's use the ones installed by the recently
introduced host-gnuconfig package.

Note that this requires using CONFIG_UPDATE_HOOK as a
PRE_CONFIGURE_HOOKS rather than a POST_PATCH_HOOKS. Indeed,
dependencies in <pkg>_DEPENDENCIES are only ready for the configure
step, not the patch step.

We have considered using <pkg>_PATCH_DEPENDENCIES, but it doesn't
work, because such dependencies are only on the "patch" step of the
depended package, so host-gnuconfig would not be installed.

However, semantically speaking, CONFIG_UPDATE_HOOK,
LIBTOOL_PATCH_HOOK, AUTORECONF_HOOK and GETTEXTIZE_HOOK do not really
belong to the "configure" step: in the context of per-package out of
tree build, those steps should only be done once for all builds of a
given package. So perhaps we will need to introduce some additional
intermediate step like "prepare", between the "patch" and "configure"
steps. But for now, moving CONFIG_UPDATE_HOOK to a PRE_CONFIGURE_HOOKS
doesn't make the situation much worse, since we already have lots of
other autoconf patching going on in PRE_CONFIGURE_HOOKS.

In addition, in the autotools-package infrastructure, we ensure to not
add the host-gnuconfig to host-tar, otherwise we have a circular
dependency: host-gnuconfig being fetched from git, it creates and then
extracts a tarball, which needs host-tar. So if host-tar in turn needs
host-gnuconfig, we get a circular dependency.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>