support/download/cargo-post-process, package/pkg-cargo.mk: enable vendoring for Cargo packages
This commit adds support/download/cargo-post-process to perform the
vendoring on Cargo packages, and enables it in package/pkg-cargo.mk.

Since it changes the contents of the tarballs for ripgrep and
sentry-cli, it changes their hashes. To not have a different hash for
the same version of ripgrep and sentry-cli, we bump their versions. It
has to be done in the same commit as the Cargo vendoring to make the
series bisectable.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>